ref: Handle errors better, delete dead imports
Deletes dead imports and kills dead code. It also handles errors better
than the original method. It propagates errors up the stack until it
reaches the main function. Main is required to handle it. For now, it
unwraps and panics. Another todo is to fix that error handling as well.

[#126077553]