Fix build command
The build command was erroneously copying files in directories prefaced
by underscores because the filtering on the file was only done at the
filename level, not the path level. This commit should fix that issue
from occurring.

I also checked in the tests/target directory because it was mistakenly
exempted from VCS from the .gitignore.

[#142912325]